reorganized markets components
