+++ components/markets/index.tsx
@@ -1,8 +1,8 @@
 import { fetchFearGreedIndexData } from '@/lib/markets/fear-greed-index';
 import { fetchAllTickerData } from '@/lib/markets/tickers';
-import { FearGreedIndex } from '@/components/fear-greed-index';
-import StockTickers from '@/components/stock-tickers';
-import CryptoTickers from '@/components/crypto-tickers';
+import { FearGreedIndex } from '@/components/markets/fear-greed-index';
+import StockTickers from '@/components/markets/stock-tickers';
+import CryptoTickers from '@/components/markets/crypto-tickers';
 
 export default async function Markets() {
   const [fearGreedIndexData, tickerData] = await Promise.all([
